+++ src/requests/users-req.js
@@ -3,7 +3,7 @@
 */
 export async function saveUser (obj) {
     let data;
-    await fetch('http://localhost:8080/auth/register', {
+    await fetch('http://localhost:8080/user/register', {
         method: 'POST',
         headers: {
             'Content-Type': 'application/json'
@@ -20,7 +20,7 @@ export async function saveUser (obj) {
 */
 export async function login (obj) {
     let data;
-    await fetch('http://localhost:8080/auth/login', {
+    await fetch('http://localhost:8080/user/login', {
         method: 'POST',
         headers: {
             'Content-Type': 'application/json'
@@ -33,11 +33,29 @@ export async function login (obj) {
 }
 
 /*
+    Should allow us to log in a user (doesn't really do that rn Jan12)
+*/
+export async function getUserBusiness (obj) {
+    let data;
+    await fetch('http://localhost:8080/user/business', {
+        method: 'POST',
+        headers: {
+            'Content-Type': 'application/json'
+        },
+        body: JSON.stringify(obj)
+    })
+    .then(res => data = res.json())
+    .catch(err => console.log(err));   
+    return data;
+}
+
+/*
+    TODO
     Allows us to update a users password
 */
 export async function updatePassword(obj){
     let data;
-    await fetch('http://localhost:8080/auth/password', {
+    await fetch('http://localhost:8080/user/', {
         method: 'POST',
         headers: {
             'Content-Type': 'application/json'