+++ src/CHANGELOG.md
@@ -6,6 +6,7 @@ Create a new version whenever we update main/master.
 Add items closer to the version number so the date go from newest to oldest (top to bottom).
 
 ### *Version 0.1.6*
+ - 2022/01/12 - [Build CI](https://sd-onlinepos.atlassian.net/browse/ON-39) 
  - 2022/01/09 - [Header Changes to support before and after login](https://sd-onlinepos.atlassian.net/browse/ON-75) 
  - 2022/11/28 - [Create Reuseable components](https://sd-onlinepos.atlassian.net/browse/ON-28) 
  - 2022/11/28 - [Create Layout for Landing Page](https://sd-onlinepos.atlassian.net/browse/ON-41)