+++ src/CHANGELOG.md
@@ -6,6 +6,7 @@ Create a new version whenever we update main/master.
 Add items closer to the version number so the date go from newest to oldest (top to bottom).
 
 ### *Version 0.1.6*
+ - 2023/01/15 - [Adds data models and fetching and posting to them](https://sd-onlinepos.atlassian.net/browse/ON-19)
  - 2023/01/12 - [Create Layout for Create Business Page](https://sd-onlinepos.atlassian.net/browse/ON-73) 
  - 2023/01/12 - [Build CI](https://sd-onlinepos.atlassian.net/browse/ON-39) 
  - 2023/01/09 - [Header Changes to support before and after login](https://sd-onlinepos.atlassian.net/browse/ON-75) 
@@ -13,4 +14,3 @@ Add items closer to the version number so the date go from newest to oldest (top
  - 2022/11/28 - [Create Layout for Landing Page](https://sd-onlinepos.atlassian.net/browse/ON-41) 
  - 2022/11/20 - [Create Layout for Create Account Page](https://sd-onlinepos.atlassian.net/browse/ON-45) 
  - 2022/11/20 - [Create CHANGELOG.md file](https://sd-onlinepos.atlassian.net/browse/ON-40)
-