+++ src/CHANGELOG.md
@@ -5,9 +5,11 @@ Create a new version whenever we update main/master.
 
 Add items closer to the version number so the date go from newest to oldest (top to bottom).
 
-### *Version 0.1.6*
+### *Version 0.1.8*
+ - 2023/01/15 - [Implement Password Hashing](https://sd-onlinepos.atlassian.net/browse/ON-83)
+ - 2023/01/15 - [Configure Authentication Middleware](https://sd-onlinepos.atlassian.net/browse/ON-43)
  - 2023/01/15 - [Adds data models and fetching and posting to them](https://sd-onlinepos.atlassian.net/browse/ON-19)
- - 2023/01/12 - [Create Layout for Create Business Page](https://sd-onlinepos.atlassian.net/browse/ON-73) 
+ - 2023/01/12 - [Create Layout for Create Business Page](https://sd-onlinepos.atlassian.net/browse/ON-73)
  - 2023/01/12 - [Build CI](https://sd-onlinepos.atlassian.net/browse/ON-39) 
  - 2023/01/09 - [Header Changes to support before and after login](https://sd-onlinepos.atlassian.net/browse/ON-75) 
  - 2022/11/28 - [Create Reuseable components](https://sd-onlinepos.atlassian.net/browse/ON-28)